pack llama_gemm
